45_RPM_Spinning_Record_Animation/requirements.txt
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
